NativeMessageHandlerService refactor and additional types
