+++ apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/credentialRetrievePayload.ts
@@ -1,0 +1,4 @@
+export type CredentialRetrievePayload = {
+  userId: string;
+  uri: string;
+};
