+++ apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/passwordGeneratePayload.ts
@@ -1,0 +1,3 @@
+export type PasswordGeneratePayload = {
+  userId: string;
+};
