+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/accountStatusResponse.ts
@@ -1,0 +1,6 @@
+export type AccountStatusResponse = {
+  id: string;
+  email: string;
+  status: "locked" | "unlocked";
+  active: boolean;
+};
