+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/cannotDecryptErrorResponse.ts
@@ -1,0 +1,3 @@
+export type CannotDecryptErrorResponse = {
+  error: "cannot-decrypt";
+};
