+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/failureStatusResponse.ts
@@ -1,0 +1,3 @@
+export type FailureStatusResponse = {
+  status: "failure";
+};
