+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/generateResponse.ts
@@ -1,0 +1,3 @@
+export type GenerateResponse = {
+  password: string;
+};
