+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/successStatusResponse.ts
@@ -1,0 +1,3 @@
+export type SuccessStatusResponse = {
+  status: "success";
+};
