+++ apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/userStatusErrorResponse.ts
@@ -1,0 +1,3 @@
+export type UserStatusErrorResponse = {
+  error: "locked" | "not-active-user";
+};
