+++ apps/desktop/src/models/nativeMessaging/index.ts
@@ -1,6 +1,17 @@
-export * from "./cipherResponse";
-export * from "./credentialCreatePayload";
-export * from "./credentialUpdatePayload";
+export * from "./encryptedMessagePayloads/credentialCreatePayload";
+export * from "./encryptedMessagePayloads/credentialRetrievePayload";
+export * from "./encryptedMessagePayloads/credentialUpdatePayload";
+export * from "./encryptedMessagePayloads/passwordGeneratePayload";
+
+export * from "./encryptedMessageResponses/accountStatusResponse";
+export * from "./encryptedMessageResponses/cannotDecryptErrorResponse";
+export * from "./encryptedMessageResponses/cipherResponse";
+export * from "./encryptedMessageResponses/encryptedMessageResponse";
+export * from "./encryptedMessageResponses/failureStatusResponse";
+export * from "./encryptedMessageResponses/generateResponse";
+export * from "./encryptedMessageResponses/successStatusResponse";
+export * from "./encryptedMessageResponses/userStatusErrorResponse";
+
 export * from "./decryptedCommandData";
 export * from "./encryptedCommand";
 export * from "./encryptedMessage";
